feat: add story selection
* Added the ability to allow the user to select a story.

In stories.py:
* Added a new story and put all of the stories in a List, so that the new
feature can be used and tested.

In app.py:
* Moved the homepage's form and logic to a new "/story/prompt" route, in
correspondence to instructions.
* Moved the generated story route to "/story/<story_id>/result", in
correspondence to instructions.
* The generated story page uses an ID in the URL for organization and to
better describe that the generated story belongs to a particular story
ID.
* The prompt page doesn't have a story ID in the URL, because it didn't
seem possible to get the ID from the dropdown on the homepage and apply
it to the form action on the homepage without JavaScript.  If
JavaScript will be implemented, the prompt page URL can be changed to
include the ID.  The prompt page will be using a query to determine
story ID.

In home.html:
* Replaced the HTML code with a form and a dropdown to select a Mad Libs
story, as instructed.

In prompt.html:
* Moved the HTML code from "home.html" to here.
* Changed the form action location to reflect the new URL design.